Cleanup + rollback avoidance (#65)
* Avoid ignoring hidden files

Hidden files should generally not be generated by build scripts, so
they do not need to be ignored.  And some hidden files
(e.g. .gitignore) are quite useful to keep under revision control.

If there are specific hidden files that do need to be ignored, they
should be added individually.

* Ignore emacs backup files.

When emacs edits foo.txt, it saves a temporary backup file named
foo.txt~ -- we never want to commit those to revision control.

* Add coding convention guidance for editing policy.json with emacs

These instructions keep policy.json well-structured if anyone edits it
with emacs, and will be ignored by people using other editors.

* Fix link to MTA-STS, add links to MTA-STS and TLSRPT.

We probably want to point people to the latest drafts of MTA-STS and
TLSRPT, rather than to a static version, since we don't want the
starttls policy list tied to a version that has been obsoleted.

* Store authority key directly, and move from gpg to gpgv.

We were storing the fingerprint and the signing subkey long key ID,
and then fetching the key from the keyservers and importing it into
the superuser's GnuPG keyring.

It makes more sense to just store the key itself directly, and not
need to worry about potential keyserver problems. This also avoids
tampering with the superuser's keyring.

Furthermore, gpgv is a simpler tool than gpg itself, and is more
narrowly-tailored for signature verification.  So we use the simplest
tool that gets the job done.

Additionally, we explicitly specify both the signature file and the
file that is to be verified, to avoid a common trap where the
signature file itself contains a signed message, leaving the external
file essentially ignored.

Note that the key is stored here in simple base64 format, rather than
as an OpenPGP "ascii-armored" blob, because gpgv prefers its keyrings
in raw binary form.

I considered placing the key in a separate file, but then you need the
script to know where the other file is, and that seemed like more
hassle than it was worth.

If the key itself gets updated in the future, or if you switch keys,
you can update them here, using something like the following to
generate the base64 encoded data:

    gpg --export-options export-minimal \
      --export-filter 'drop-subkey=usage!~s'
      --export B693F33372E965D76D55368616EEA65D03326C9D \
      | base64

* Ensure that the signature does not roll back.

If we've never fetched a version of the policy, we accept any
signature.

But if we have a more recent signed version of the policy, we should
refuse any attempt to pass off an older signature than the most recent
one we've already seen.