README rehaul (#91)
* README update

Updates to the main README in the repository. Specifies new goals,
relative to ecosystem changes since 2014. In particular, STARTTLS
is widely adopted now, but is still susceptible to active network
attackers (via simple downgrade, DNS MitM, or TLS MitM).

Given those things, here is an update to the README that reflects
the new goals of the project relative to the current status of
the email ecosystem.

* Remove min-tls-version setting.

When this field was initially proposed, many mailservers had not yet made the move to
TLSv1.2. We're at a point where TLSv1.2 support is nearly universal, and should be
a security prerequisite. Also, the goal of this project does not include protecting against
TLS version downgrade attacks.